Add files via upload
the task was to detect fraudulent transactions in a financial dataset. The first step was to transform the string values in the columns nameOrig and nameDest by replacing specific substrings with numerical values 9 and 7. This conversion was done to preserve the essence of the column while enabling numerical processing. As encoding was not feasible due to the large size of the dataset, this alternative solution was adopted. During the data exploration stage, a single outlier was detected, but it was suspected to be a fraud case. Further analysis revealed that only the Transfer and Cash-out categories in the type column were involved in fraud. The heatmap was used to identify the existence of multicollinearity between oldbalanceOrg and newbalanceOrg, as well as between oldbalanceDest and newbalanceDest. To mitigate this issue, Principal Component Analysis (PCA) was employed to reduce the weight of these highly correlated columns. As class 1, which represents fraud cases, was a minority class in the dataset, Synthetic Minority Over-sampling Technique (SMOTE) was used to balance the class distribution. Finally, the logistic regression model was trained and achieved an accuracy of 95%.   You can find the dataset : https://www.kaggle.com/code/benroshan/transaction-fraud-detection/data