Add files via upload
A customer information dataset was obtained with demographic and
behavior data, as well as churn status. Data cleaning and
preprocessing were done to handle missing data and convert categorical variables. Feature selection was performed to choose
relevant features for the model. A predictive model was built using machine learning algorithms, such as Logistic Regression, Random
Forest or Gradient Boosting. The model was trained and evaluated on a portion of the data, with performance metrics such as accuracy, precision, recall, F1-score and AUC-ROC being used to measure its
success.